Gracefully handle cases where gcov output has 100% coverage of files
The code assumed that there would always be some lines missing from
the gcovr data.  However, if all a file's executable code was covered,
the line was blank following the "100%".  The attempt to read a number
from a string that contained no digits caused an exception to be
thrown.  The code has been modified to handle those cases where there
are not missing lines gracefully.